split api layer based on functionality
